Adder example for adding fixed point number
The entity adder_fix_pt in demo_ieee_proposed.vhd takes in an array of 'p' fixed point numbers of the form sfixed(n+1 downto -m) and outputs their sum in the form sfixed(n+1 downto -m) . The entity named demo_ieee_proposed is actually a wrapper for adder_fix_pt so as to provide an interface to testbench only in std_logic_vectors. This is done because modelsim talks only in the language of std_logic_vectors. :D